% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census.names}
\alias{census.names}
\title{Frequency of name occurrence on US birth certificates.}
\format{
A data frame with 164850 rows and 5 variables:
\describe{
\item{name}{names listed on US birth certificates}
\item{first_name_value}{count of times the name is listed on a birth certificate as a first name}
\item{last_name_value}{count of times the name is listed on a birth certificate as a last name}
}
}
\source{
\url{https://www.census.gov/en.html}
}
\usage{
census.names
}
\description{
A dataset containing names cross-tabbed by whether they
are listed as first or last names in the US Census.
}
\keyword{datasets}
