% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.gender.R
\name{determine.gender}
\alias{determine.gender}
\title{Determine gender from name matrix}
\usage{
determine.gender(x)
}
\arguments{
\item{x}{Provide the character vector to parse.}
}
\description{
This function returns
}
\examples{
x <- 'livingston III,  Mr. MICHAEL JOHN9'
x <- prep.name(x)
x <- strsplit(x,' ')[[1]]
cd <- get.census.data(x)
determine.gender(cd)
}
\keyword{gener}
