% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine.surname.R
\name{determine.surname}
\alias{determine.surname}
\title{Determine the Surname (or Last Name)}
\usage{
determine.surname(x)
}
\arguments{
\item{x}{Provide the 'census matrix' that is to be evaluated.  See examples.}
}
\description{
This function uses results from \code{\link[name.parser]{get.census.matrix}} to evaluate and derive the surname of any given character vector of Names.
}
\note{
The function returns a two element integer vector.  The first element is the ordinal of the recommend/suggested
surname.  The second element is the confidence of the recommendation.
}
\examples{
x <- 'livingston III,  Mr. MICHAEL JOHN9'
x <- prep.name(x)
x <- strsplit(x,' ')[[1]]
cd <- get.census.data(x)
print(x)
determine.surname(cd)
}
\author{
mjfii \href{mailto:mick.flanigan@gmail.com}{mick.flanigan@gmail.com}
}
\keyword{matrix}
\keyword{surname}
