% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.census.data.R
\name{get.census.data}
\alias{get.census.data}
\title{Pull Name Matrix bases on character vector}
\usage{
get.census.data(x)
}
\arguments{
\item{x}{Provide the name character vector to parse.}
}
\value{
data.table
}
\description{
This function returns
}
\note{
The return data.table is sorted based on the ordinals of the character vector arguement.
}
\examples{
x <- 'livingston III,  Mr. MICHAEL JOHN9'
x <- prep.name(x)
x <- strsplit(x,' ')[[1]]
get.census.data(x)
}
\keyword{census}
\keyword{name}
