% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.example.names.R
\name{get_example_names}
\alias{get_example_names}
\title{Build a test sample of names from IRS 990s}
\usage{
get_example_names(n = 10)
}
\arguments{
\item{n}{Size of the sample vector of names (maximum 1000).}
}
\description{
This function builds a test vector of up to 1000 names.
}
\examples{
x <- get_example_names( n=10 )
pn <- parse.names( x )
head( pn )

# micro-benchmark performance
x <- get_example_names( n=1000 )
start_time <- Sys.time()
pn <- parse.names( x )
end_time <- Sys.time()
end_time - start_time
}
\keyword{test}
