% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{names.100}
\alias{names.100}
\title{Sample of 100 names for testing purposes}
\format{
An object of class \code{character} of length 100.
}
\usage{
names.100
}
\description{
Sample of 100 names for testing purposes
}
\keyword{datasets}
