% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{names.1000}
\alias{names.1000}
\title{Sample of 1,000 names for testing purposes}
\format{
An object of class \code{character} of length 1004.
}
\usage{
names.1000
}
\description{
Sample of 1,000 names for testing purposes
}
\keyword{datasets}
