% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.name.R
\name{parse.name}
\alias{parse.name}
\title{Parse single name}
\usage{
parse.name(x, prefixes = prx, suffixes = sfx)
}
\arguments{
\item{x}{Provide the character vector to parse.}

\item{prefixes}{x}

\item{suffixes}{x}
}
\description{
This function returns
}
\examples{
x <- 'LIVINGSTON III MICHAEL JOHN'
parse.name(x)
}
\keyword{remove}
