% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.names.R
\name{parse.names}
\alias{parse.names}
\title{Parse a Set of Names}
\usage{
parse.names(x, ...)
}
\arguments{
\item{x}{Provide the charater vector to parse.}

\item{...}{Alternate arguments for suffixes, prefixes, etc.}
}
\description{
This function returns a data frame of parse attributes based on an input of credit card names.
}
\examples{
x <- 'livingston III,  Mr. MICHAEL JOHN9'
parse.names(x)
}
\keyword{parse}
