% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.name.R
\name{prep.name}
\alias{prep.name}
\title{Prep Character Vector for Name Parsing}
\usage{
prep.name(x)
}
\arguments{
\item{x}{Provide the character vector to prep.}
}
\description{
This function
}
\note{
If more then one element is passed in, only the first element will be preppred and returned.
}
\examples{
x <- 'livingston III,   MICHAEL JOHN9'
prep.name(x)
}
