% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{prx}
\alias{prx}
\title{List of known prefix strings}
\format{
An object of class \code{character} of length 60.
}
\usage{
prx
}
\description{
List of known prefix strings
}
\keyword{datasets}
