% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.elements.R
\name{remove.elements}
\alias{remove.elements}
\title{Remove Character Elements From a Vector}
\usage{
remove.elements(x, remove.this)
}
\arguments{
\item{x}{Provide the character vector to parse.}

\item{remove.this}{Provide the elements that you want to remove.}
}
\description{
This function returns a two part list of character vectors.  First, the new vector with the removed strings,
and second, the vector with the elements that have been removed.
}
\examples{
x <- c('MICHAEL','JOHN','LIVINGSTON','III')
remove.elements(x, c('JR','II','III','IV'))
}
\keyword{remove}
